endian

+++ include/CDScene/IO/OutputArchive.hpp
@@ -49,8 +49,8 @@ public:
 			bufferBytes = sourceBufferBytes;
 		}
 
-		m_pOStream->write(reinterpret_cast<const char*>(&bufferBytes), sizeof(uint64_t));
-		m_pOStream->write(reinterpret_cast<const char*>(data), sourceBufferBytes);
+		m_pOStream->write(reinterpret_cast<const char *>(&bufferBytes), sizeof(uint64_t));
+		m_pOStream->write(reinterpret_cast<const char *>(data), sourceBufferBytes);
 
 		return *this;
 	}
@@ -63,7 +63,6 @@ private:
 		{
 			if constexpr (SwapBytesOrder)
 			{
-				// T checkedData = std::byteswap(data);
 				T checkedData = byte_swap<T>(data);
 				m_pOStream->write(reinterpret_cast<const char*>(&checkedData), sizeof(T));
 			}
@@ -78,13 +77,11 @@ private:
 			{
 				if constexpr (4 == sizeof(T))
 				{
-					// uint32_t checkedData = std::byteswap(static_cast<uint32_t>(data));
 					float checkedData = byte_swap<float>(data);
 					m_pOStream->write(reinterpret_cast<const char*>(&checkedData), sizeof(T));
 				}
 				else if constexpr (8 == sizeof(T))
 				{
-					// uint64_t checkedData = std::byteswap(static_cast<uint64_t>(data));
 					double checkedData = byte_swap<double>(data);
 					m_pOStream->write(reinterpret_cast<const char*>(&checkedData), sizeof(T));
 				}
@@ -100,15 +97,14 @@ private:
 		}
 		else if constexpr (std::is_same<T, std::string>())
 		{
-			size_t dataLength = data.size();
+			uint64_t dataLength = data.size();
 			if constexpr (SwapBytesOrder)
 			{
 				// std::string is just array of 1 byte char so don't need to swap bytes.
-				// dataLength = std::byteswap(dataLength);
-				dataLength = byte_swap<size_t>(dataLength);
+				dataLength = byte_swap<uint64_t>(dataLength);
 			}
 
-			m_pOStream->write(reinterpret_cast<const char*>(&dataLength), sizeof(size_t));
+			m_pOStream->write(reinterpret_cast<const char*>(&dataLength), sizeof(uint64_t));
 			m_pOStream->write(data.c_str(), data.size());
 		}
 		else
